# src/backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LensIQ Multi-Source CCTV Backend & Streaming Gateway API",
    "description": "Enterprise CCTV monitoring REST API and Streaming Gateway supporting multiple video source types (RTSP and Hikvision P2P) with zero secret leakage, single-pipeline multiplexing, and AI stream normalization.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local Development Server"
    },
    {
      "url": "https://stream.lensiq.cloud/api/v1",
      "description": "Production Streaming Gateway"
    }
  ],
  "paths": {
...
    "/cameras": {
      "get": {
        "summary": "List cameras with multi-tenant filtering",
        "parameters": [
          { "name": "companyId", "in": "query", "schema": { "type": "string" } },
          { "name": "branchId", "in": "query", "schema": { "type": "string" } },
          { "name": "sourceType", "in": "query", "schema": { "type": "string", "enum": ["rtsp", "hikvision_p2p"] } }
        ],
        "responses": {
          "200": {
            "description": "List of safe camera DTOs (zero credentials exposed)"
          }
        }
      },
      "post": {
        "summary": "Create a new camera (RTSP or Hikvision P2P)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCameraRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Camera successfully created and registered with credentials vault"
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateCameraRequest": {
        "type": "object",
        "required": ["name", "company_id", "branch_id", "source_type"],
        "properties": {
          "name": { "type": "string", "example": "Cashier 01" },
          "company_id": { "type": "string", "format": "uuid" },
          "brand_id": { "type": "string", "format": "uuid" },
          "branch_id": { "type": "string", "format": "uuid" },
          "source_type": { "type": "string", "enum": ["rtsp", "hikvision_p2p"] },
          "location_description": { "type": "string", "example": "Counter 1 - Checkout" },
          "rtsp_url": { "type": "string", "example": "rtsp://camera.local/live/ch1" },
          "hik_device_id": { "type": "string", "example": "HIK-DEV-9011" },
          "hik_serial_number": { "type": "string", "example": "SER123456" },
          "hik_channel": { "type": "integer", "example": 1 },
          "credentials_payload": {
            "type": "object",
            "description": "Sensitive credentials securely registered into vault; never stored plain in database"
          },
          "stream_profile": { "type": "string", "enum": ["main", "sub"], "default": "main" }
        }
      },
...
    }
  }
}